requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
blobfile==3.0.0
camel-converter==4.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==4.4.2
deep-translator==1.11.4
distro==1.9.0
einops==0.8.0
elastic-transport==8.15.1
elasticsearch==8.15.1
exceptiongroup==1.2.2
fairscale==0.4.0
faiss-cpu==1.9.0
faiss-gpu==1.7.2
fastapi==0.115.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-huggingface==0.1.0
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langsmith==0.1.135
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.2
meilisearch==0.31.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy==1.12.0
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai==1.52.0
openai-whisper==20240930
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.2
pyarrow==17.0.0
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyDeprecate==0.3.2
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.2.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.35
starlette==0.39.2
sympy==1.13.3
tenacity==8.5.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==0.4.12
tokenizers==0.20.1
tomli==2.0.2
torch==2.4.1
torchmetrics==0.7.3
torchscale==0.2.0
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
wcwidth==0.2.13
yarl==1.15.2
zipp==3.20.2
